+++ quickcheck-plutus-instances/CHANGELOG.md
@@ -9,10 +9,12 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 ### Added
 
 * Instances of `Arbitrary`, `CoArbitrary` and `Function` for
-  `Plutus.V1.Value.Value`.
+  `Value`, `Address`, `Credential`, `TxOut`, `StakingCredential`. 
 * Instances of `Arbitrary`, `Arbitrary1`, `Arbitrary2`, `CoArbitrary` and
   `Function` for `PlutusTx.AssocMap.Map`.
 * `UniqueMap` wrapper around `AssocMap.Map` which ensures unique keys.
+* `UniqueList` wrapper around lists to ensure unique values.
+* `NonNegative` wrapper for generating non-negative (Plutus) numerical values.
 
 ### Changed